test: 🔄 preserveNodeMaps: move to a separate file
